feat(finance): add payment entity and wallet transaction system

+++ src/main/java/ir/maktabsharif138/home_service_system/dto/request/HomeServiceCreateRequest.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class HomeServiceCreateRequest {
     @NotBlank(message = "Service name is required")
@@ -14,7 +16,7 @@ public class HomeServiceCreateRequest {
 
     @NotNull(message = "Base price is required")
     @Positive(message = "Base price must be positive")
-    private Double basePrice;
+    private BigDecimal basePrice;
 
     private Long parentServiceId;
 }
