+++ src/main/java/ir/maktabsharif138/home_service_system/dto/request/OrderCreateRequest.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,7 +15,7 @@ public class OrderCreateRequest {
 
     @NotNull(message = "Proposed price is required")
     @Positive(message = "Proposed price must be positive")
-    private Double proposedPrice;
+    private BigDecimal proposedPrice;
 
     @NotNull(message = "Start date and time is required")
     @Future(message = "Start date must be in the future")
