+++ src/main/java/ir/maktabsharif138/home_service_system/dto/response/ExpertOrderHistoryResponse.java
@@ -3,6 +3,7 @@ package ir.maktabsharif138.home_service_system.dto.response;
 import ir.maktabsharif138.home_service_system.entity.enums.OrderStatus;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +11,7 @@ public class ExpertOrderHistoryResponse {
 
     private Long orderId;
     private String description;
-    private Double proposedPrice;
+    private BigDecimal proposedPrice;
     private LocalDateTime orderDate;
     private OrderStatus orderStatus;
     private String customerName;
