+++ src/main/java/ir/maktabsharif138/home_service_system/dto/response/HomeServiceResponse.java
@@ -1,6 +1,8 @@
 package ir.maktabsharif138.home_service_system.dto.response;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -8,7 +10,7 @@ public class HomeServiceResponse {
     private Long id;
     private String name;
     private String description;
-    private Double basePrice;
+    private BigDecimal basePrice;
     private Long parentServiceId;
     private List<HomeServiceResponse> subServices;
 }
