+++ src/main/java/ir/maktabsharif138/home_service_system/dto/response/OfferResponse.java
@@ -2,12 +2,14 @@ package ir.maktabsharif138.home_service_system.dto.response;
 
 import ir.maktabsharif138.home_service_system.entity.enums.OfferStatus;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class OfferResponse {
     private Long id;
-    private Double proposedPrice;
+    private BigDecimal proposedPrice;
     private LocalDateTime proposedStartTime;
     private Integer durationHours;
     private LocalDateTime offerDate;
